// src/enums/Redemptions.ts
export enum RedemptionIds {
  GetVip = "ddcb6566-6b02-447c-a396-6b93ade3b480",
  Hidrate = "3dfb7ed4-b696-4dff-92d7-18f9bfa3dd4a",
  HighlightMessage = "df93120a-9ebe-493c-91e1-42338400dc35",
  KaraokeTime = "53c72f19-61ae-4a27-bbe0-c8100e2fce8e",
  LightTheme2m = "f60151fc-7f62-48c6-b225-21007e0dd544",
  LightTheme5m = "a7c10cae-dbea-4bdc-aeb5-ebd3d0f77b93",
  RussianRoulette = "c38a6de3-ae92-4add-bd56-c6346c19ea93",
  StealVip = "84283c28-a2c2-4da6-99c2-b662417b9d49",
  TimeoutFriend = "1a6bd16d-2882-4125-9f8a-e3c7194ea8b6",
}